Voice thread nearly finished

+++ app/src/main/java/com/nevermore/walkietalkie/VoiceChannel.java
@@ -1,6 +1,6 @@
 package com.nevermore.walkietalkie;
 
 public class VoiceChannel extends Channel {
-    int state = 0;
+    byte state = 0;
     String speaker;
 }
